requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
PyMuPDF>=1.22.3
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
langchain>=0.0.200
google-cloud-aiplatform>=1.35.0
langchain-google-genai>=0.0.1
python-dotenv>=0.19.0
pytest>=7.0.0
black>=22.3.0
flake8>=4.0.0